fix bug: SEQUOIADBMAINSTREAM-2082 only delete full file
git-svn-id: http://192.168.20.11/sequoiadb/trunk@26008 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/tools/replay/rplReplayer.cpp
@@ -379,7 +379,8 @@ namespace replay
       PD_LOG(PDEVENT, "Replay archive log file[%s] successfully",
              file.c_str());
 
-      if (_options->remove())
+      // only delete full file
+      if (_options->remove() && _archiveFileMgr.isFullFileName(file))
       {
          rc = ossFile::deleteFile(file);
          if (SDB_OK != rc)
